+++ source/InjectHooksMain.cpp
@@ -139,10 +139,10 @@
 #include "InteriorManager_c.h"
 #include "Checkpoint.h"
 #include "Checkpoints.h"
-#include "IKChain_c.h"
-#include "IKChainManager_c.h"
-#include "BoneNode_c.h"
-#include "BoneNodeManager_c.h"
+#include "Ragdoll/IKChain.h"
+#include "Ragdoll/IKChainManager.h"
+#include "Ragdoll/BoneNode.h"
+#include "Ragdoll/BoneNodeManager.h"
 #include "TheScripts.h"
 #include "RunningScript.h"
 #include "Scripted2dEffects.h"
