+++ source/game_sa/Conversations/PedToPlayerConversations.cpp
@@ -1,6 +1,6 @@
 #include "PedToPlayerConversations.h"
 
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 
 void CPedToPlayerConversations::InjectHooks() {
     RH_ScopedClass(CPedToPlayerConversations);
