+++ source/game_sa/Events/EventEditableResponse.cpp
@@ -3,7 +3,7 @@
 #include "EventEditableResponse.h"
 
 #include "PedType.h"
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 
 void CEventEditableResponse::InjectHooks() {
     RH_ScopedVirtualClass(CEventEditableResponse, 0x85AB80, 17);
