+++ source/game_sa/Ragdoll/BoneInfo.h
