+++ source/game_sa/Ragdoll/BoneNode.h
