+++ source/game_sa/Tasks/TaskTypes/TaskComplexEnterCar.cpp
@@ -1,6 +1,6 @@
 #include "StdInc.h"
 
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 
 #include "TaskSimpleBikeJacked.h"
 #include "TaskComplexEnterCar.h"
