+++ source/game_sa/Tasks/TaskTypes/TaskComplexFleeEntity.cpp
@@ -1,7 +1,7 @@
 #include "StdInc.h"
 #include "TaskComplexFleeEntity.h"
 #include "TaskComplexFleePoint.h"
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 #include "InterestingEvents.h"
 #include "EventSeenPanickedPed.h"
