+++ source/game_sa/Tasks/TaskTypes/TaskComplexInvestigateDeadPed.cpp
@@ -1,5 +1,5 @@
 #include "StdInc.h"
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 #include "TaskComplexInvestigateDeadPed.h"
 #include "TaskComplexGoToPointAndStandStill.h"
 #include "TaskSimpleStandStill.h"
