+++ source/game_sa/Tasks/TaskTypes/TaskSimpleIKLookAt.cpp
@@ -1,5 +1,5 @@
 #include "StdInc.h"
-#include "IKChainManager_c.h"
+#include "Ragdoll/IKChainManager.h"
 #include "TaskSimpleIKLookAt.h"
 
 void CTaskSimpleIKLookAt::InjectHooks() {
